refactor portfolio book routes
